dataoracle/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pymongo==4.8.0
pycaret==3.3.2
scikit-learn==1.5.2
xgboost==2.1.1
lightgbm==4.5.0
shap==0.46.0
pandas==2.2.3
numpy==2.1.1
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
langchain==0.3.0
openai==1.47.0
mlflow==2.17.0
python-multipart==0.0.9
websockets==12.0
python-dotenv==1.0.1